refactor: using stateless component/PureComponent
Stateless component can be declared as a function that has no state and
do not have the component lifecycle methods.

PureComponent gives a considerable  increase in performance because it
reduces the number of render operation in the application.
This is because it changes the lifecycle method shouldComponentUpdate
and adds some logic to automatically check whether a re-render is
required for component.